Fixed packet corruption
Issue with packet corruption found to be caused by meddling with the
calls to md5 hashing that caused the bytes that were fed into md5 buffer
to be overwritten with the hash itself. So when the packet was passed to
peer.process(), when packet.Owner() was called, the hashing would
overwrite the bytes in the packet and screw all the things.

Extracted the hashing to a separate file/function to instead copy the
bytes rather than meddle with the raw ones.